refactor: consolidate transfer from scenarios into one file
